work on serial number

+++ components/dlms/dlms.cpp
@@ -246,7 +246,7 @@ namespace esphome {
       //this->s_serial_number_->publish_state("test");
       uint32_t serial_number;
       serial_number = (iv[4] & 0xf) << 24 | iv[4] << 16 | iv[5] << 8 | iv[7];
-      ESP_LOGI(TAG, "SML Data serial numbers: %i", (size_t) serial_number);
+      ESP_LOGI(TAG, "SML Data serial numbers: %i", (uint32_t) serial_number);
       //uint32_t serial_number = (iv[4] & 0xf) << 24 | iv[5] << 16 | iv[6] << 8 | iv[7];
       //ESP_LOGI(TAG, "SML Data serial numberi: %i", serial_number);
       //ESP_LOGI(TAG, "SML Data serial numberd: %d", serial_number);
